requirements.txt
flask
flask-restful
numpy
pillow
httpretty
requests